Changed group-id and bundle-name. Added description to service targeting the review comments.
