+++ nifi-nar-bundles/nifi-adx-bundle/nifi-adx-api/pom.xml
@@ -17,9 +17,9 @@
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
-        <groupId>org.sebkenter.nifi</groupId>
-        <artifactId>adx-processor</artifactId>
-        <version>1.0</version>
+        <groupId>org.apache.nifi</groupId>
+        <artifactId>nifi-adx-bundle</artifactId>
+        <version>1.14.0-SNAPSHOT</version>
     </parent>
 
     <artifactId>nifi-adx-api</artifactId>
