+++ nifi-nar-bundles/nifi-adx-bundle/nifi-adx-nar/pom.xml
@@ -17,13 +17,13 @@
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
-        <groupId>org.sebkenter.nifi</groupId>
-        <artifactId>adx-processor</artifactId>
-        <version>1.0</version>
+        <groupId>org.apache.nifi</groupId>
+        <artifactId>nifi-adx-bundle</artifactId>
+        <version>1.14.0-SNAPSHOT</version>
     </parent>
 
     <artifactId>nifi-adx-nar</artifactId>
-    <version>1.0</version>
+    <version>1.14.0-SNAPSHOT</version>
     <packaging>nar</packaging>
     <properties>
         <maven.javadoc.skip>true</maven.javadoc.skip>
@@ -32,9 +32,9 @@
 
     <dependencies>
         <dependency>
-            <groupId>org.sebkenter.nifi</groupId>
+            <groupId>org.apache.nifi</groupId>
             <artifactId>nifi-adx-processors</artifactId>
-            <version>1.0</version>
+            <version>1.14.0-SNAPSHOT</version>
         </dependency>
     </dependencies>
