+++ nifi-nar-bundles/nifi-adx-bundle/nifi-adx-processors/src/main/resources/META-INF/services/org.apache.nifi.processor.Processor
@@ -12,4 +12,4 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-org.sebkenter.nifi.processors.adx.AzureAdxIngestProcessor
+org.apache.nifi.processors.adx.AzureAdxIngestProcessor
