+++ nifi-nar-bundles/nifi-adx-bundle/nifi-adx/pom.xml
@@ -17,9 +17,9 @@
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
-        <groupId>org.sebkenter.nifi</groupId>
-        <artifactId>adx-processor</artifactId>
-        <version>1.0</version>
+        <groupId>org.apache.nifi</groupId>
+        <artifactId>nifi-adx-bundle</artifactId>
+        <version>1.14.0-SNAPSHOT</version>
     </parent>
 
     <artifactId>nifi-adx</artifactId>
@@ -27,9 +27,9 @@
 
     <dependencies>
         <dependency>
-            <groupId>org.sebkenter.nifi</groupId>
+            <groupId>org.apache.nifi</groupId>
             <artifactId>nifi-adx-api</artifactId>
-            <version>1.0</version>
+            <version>1.14.0-SNAPSHOT</version>
         </dependency>
         <dependency>
             <groupId>org.apache.nifi</groupId>
