+++ nifi-nar-bundles/nifi-adx-bundle/nifi-adx/src/main/resources/META-INF/services/org.apache.nifi.controller.ControllerService
@@ -12,4 +12,4 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-org.sebkenter.nifi.adx.AzureAdxConnectionService
+org.apache.nifi.adx.AzureAdxConnectionService
