+++ nifi-nar-bundles/nifi-adx-bundle/pom.xml
@@ -19,12 +19,11 @@
     <parent>
         <groupId>org.apache.nifi</groupId>
         <artifactId>nifi-nar-bundles</artifactId>
-        <version>1.13.0</version>
+        <version>1.14.0-SNAPSHOT</version>
     </parent>
 
-    <groupId>org.sebkenter.nifi</groupId>
-    <artifactId>adx-processor</artifactId>
-    <version>1.0</version>
+    <artifactId>nifi-adx-bundle</artifactId>
+    <version>1.14.0-SNAPSHOT</version>
     <packaging>pom</packaging>
 
     <modules>
